Delete unused dependencies and resources.
Deleted unused dependencies and resources. Additionally made some
refactor and formatting. Using gradle configuration, which specifically
includes only armeabi-v7a and arm64-v8a, which should cover most
modern Android devices as suggested in TensorFlow docs.
