+++ app/build.gradle
@@ -3,13 +3,9 @@ apply plugin: 'com.google.gms.google-services'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
-apply plugin: 'com.frogermcs.androiddevmetrics'
-apply plugin: 'com.getkeepsafe.dexcount'
 apply plugin: 'com.github.ben-manes.versions'
-apply plugin: 'org.sonarqube'
 apply plugin: 'com.vanniktech.android.junit.jacoco'
 apply plugin: 'io.fabric'
-apply from: '../buildsystem/sonarqube.gradle'
 
 ext {
 
@@ -31,6 +27,9 @@ android {
         versionName '0.1.0'
         vectorDrawables.useSupportLibrary = true
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        ndk {
+            abiFilters 'armeabi-v7a', 'arm64-v8a'
+        }
     }
     buildTypes {
         debug {
@@ -149,8 +148,6 @@ configurations {
 }
 
 dependencies {
-    implementation project(':visualizer')
-
     //Kotlin
     implementation libs.kotlin
     implementation libs.constraintLayout
@@ -198,17 +195,11 @@ dependencies {
     // LeakCanary
     debugImplementation libs.leakCanary
 
-    // BlockCanary
-    debugImplementation libs.blockCanary
-
     // StrictMode
     debugImplementation libs.strictmodeNotifier
 
     // HockeyApp
     implementation libs.hockeyApp
-
-    //MaterialEditText
-    implementation libs.materialEditText
 
     //Glide
     implementation libs.glide
@@ -216,9 +207,6 @@ dependencies {
 
     //ThreeTen
     implementation libs.threeTen
-
-    //Expandable Layout
-    implementation libs.expandableLayout
 
     //Gson
     implementation libs.gson
@@ -235,12 +223,6 @@ dependencies {
 
     //Java Web Socket
     implementation libs.javaWebSocket
-
-    //Material Dialogs
-    implementation libs.materialDialogs
-
-    //ExoPlayer
-    implementation libs.exoPlayer
 
     //WebRtc
     implementation libs.googleWebRtc
