+++ app/src/debug/kotlin/co/netguru/baby/monitor/client/application/DebugMetricsHelper.kt
@@ -1,16 +1,7 @@
 package co.netguru.baby.monitor.client.application
 
 import android.content.Context
-import android.os.Handler
-import android.os.StrictMode
 import co.netguru.baby.monitor.client.application.scope.AppScope
-import com.facebook.stetho.Stetho
-import com.frogermcs.androiddevmetrics.AndroidDevMetrics
-import com.github.moduth.blockcanary.BlockCanary
-import com.github.moduth.blockcanary.BlockCanaryContext
-import com.nshmura.strictmodenotifier.StrictModeNotifier
-import com.squareup.leakcanary.LeakCanary
-import net.hockeyapp.android.CrashManager
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -31,14 +22,6 @@ import javax.inject.Inject
 class DebugMetricsHelper @Inject constructor() {
 
     internal fun init(context: Context) {
-        // LeakCanary
-        if (LeakCanary.isInAnalyzerProcess(context.applicationContext as App)) {
-            // This process is dedicated to LeakCanary for heap analysis.
-            // You should not init your app in this process.
-            return
-        }
-        LeakCanary.install(context.applicationContext as App)
-
         //Timber
         Timber.plant(Timber.DebugTree())
     }
