--- app/src/main/kotlin/co/netguru/baby/monitor/client/data/communication/firebase/FirebasePushMessage.kt
@@ -1,7 +1,0 @@
-package co.netguru.baby.monitor.client.data.communication.firebase
-
-data class FirebasePushMessage(
-        val to: String,
-        val title: String,
-        val body: String
-)
