--- app/src/main/res/layout/activity_splash.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <FrameLayout
-        android:id="@+id/splashFl"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
-</android.support.constraint.ConstraintLayout>
