+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':visualizer'
+include ':app'
