--- visualizer/build.gradle
@@ -1,46 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-
-android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
-    defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 28
-        consumerProguardFiles 'proguard-rules.pro'
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        releasePreprod {
-            debuggable true
-            minifyEnabled true
-            zipAlignEnabled true
-            useProguard true
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-        release {
-            debuggable false
-            minifyEnabled true
-            zipAlignEnabled true
-            useProguard true
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    lintOptions {
-        abortOnError false
-    }
-}
-
-dependencies {
-    //Kotlin
-    implementation libs.kotlin
-    // Unit testing
-    testImplementation test.junit
-    testImplementation test.mockito
-    testImplementation test.mockitoKotlin
-}
