--- visualizer/proguard-rules.pro
@@ -1,37 +1,0 @@
-# Add project specific ProGuard rules here.
-# By default, the flags in this file are appended to flags specified
-# in /Users/a12116/Library/Android/sdk/tools/proguard/proguard-android.txt
-# You can edit the include path and order by changing the proguardFiles
-# directive in build.gradle.
-#
-# For more details, see
-#   http://developer.android.com/guide/developing/tools/proguard.html
-
-# Add any project specific keep options here:
-
-# If your project uses WebView with JS, uncomment the following
-# and specify the fully qualified class name to the JavaScript interface
-# class:
-#-keepclassmembers class fqcn.of.javascript.interface.for.webview {
-#   public *;
-#}
--optimizationpasses 5
--dontusemixedcaseclassnames
--dontskipnonpubliclibraryclasses
--dontskipnonpubliclibraryclassmembers
--dontpreverify
--optimizations !code/simplification/arithmetic,!field/*,!class/merging/*
--keepattributes LineNumberTable,SourceFile,Signature,*Annotation*,Exceptions,InnerClasses
-
-# Keep native methods
--keepclassmembers class * {
-    native <methods>;
-}
-
-# remove log call
--assumenosideeffects class android.util.Log {
-    public static *** d(...);
-}
--assumenosideeffects class timber.log.Timber {
-    public static *** d(...);
-}
