--- visualizer/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.tyorikan.voicerecordingvisualizer"
-    android:versionCode="1"
-    android:versionName="1.0" />
